Fixes a spinner bug

+++ backend/service.sh
@@ -4,17 +4,25 @@ set -e
 PS1=fakeinteractiveshell
 source /root/.bashrc
 
+isrunning() {
+    test -f /tmp/service.pid && (ps -p $(cat /tmp/service.pid) > /dev/null)
+}
+
 startsvc() {
-    ./gradlew :launcher:installDist --console=plain
-    nohup /opt/ucloud/launcher/build/install/launcher/bin/launcher --dev --config-dir /etc/ucloud &> /tmp/service.log &
-    echo $! > /tmp/service.pid
-    sleep 0.5 # silly workaround to make sure docker exec doesn't kill us
+    if ! isrunning; then
+        ./gradlew :launcher:installDist --console=plain
+        nohup /opt/ucloud/launcher/build/install/launcher/bin/launcher --dev --config-dir /etc/ucloud &> /tmp/service.log &
+        echo $! > /tmp/service.pid
+        sleep 0.5 # silly workaround to make sure docker exec doesn't kill us
+    fi
 }
 
 stopsvc() {
-    kill $(cat /tmp/service.pid)
-    rm -f /tmp/service.pid
-    rm -f /tmp/service.log
+    if isrunning; then
+        kill $(cat /tmp/service.pid)
+        rm -f /tmp/service.pid
+        rm -f /tmp/service.log
+    fi
 }
 
 restartsvc() {
